# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: OpenLR Tool API
  description: |
    Decode and encode OpenLR location references against a self-hosted road network.

    ## Identifiers

    A segment is identified by `meta`, the reference into your source network.
    `local.roads.id` is an opaque internal key required by the OpenLR library and is
    never present in a response. `POST /encode` resolves its `path` against `meta`,
    so a decoded path can be re-encoded without translation.

    ## Status codes

    `400` means the request was at fault — a malformed code, an unknown decoder or
    encoder profile, a `meta` that matches no segment or several. A failure of the
    database or another dependency is `503`, never `400`.

    ## Only line locations

    Point-along-line, POI-with-access-point, circle, polygon, rectangle, grid and
    closed-line references are rejected. Only the binary (base64) physical format is
    supported; there is no XML endpoint.
  version: "1.1.0"
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

servers:
  - url: http://localhost:8081/api/v1
    description: Default deployment
  - url: http://localhost:3000/api/v1
    description: Through the web front end, which proxies /api to the backend

tags:
  - name: Locations
    description: Decoding and encoding OpenLR references
  - name: Network
    description: Querying the loaded road network
  - name: Operations
    description: Health, caches and configuration

paths:
  /decode:
    get:
      summary: Decode an OpenLR reference (query parameters)
      operationId: decodeGet
      tags: [Locations]
      parameters:
        - $ref: '#/components/parameters/OpenLrCode'
        - $ref: '#/components/parameters/DecoderProps'
      responses:
        '200': { $ref: '#/components/responses/DecodedPath' }
        '400': { $ref: '#/components/responses/DecodeFailed' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
    post:
      summary: Decode an OpenLR reference (JSON or form body)
      description: |
        Accepts `application/json` and `application/x-www-form-urlencoded`. The two
        forms take the same fields and behave identically.
      operationId: decodePost
      tags: [Locations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [openLrCode]
              properties:
                openLrCode:
                  type: string
                  example: CwQ1cyVM2BNWDPvw/0cTBA==
                props:
                  type: string
                  default: default
                  example: strict,relaxed,default
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [openLrCode]
              properties:
                openLrCode: { type: string }
                props: { type: string, default: default }
      responses:
        '200': { $ref: '#/components/responses/DecodedPath' }
        '400': { $ref: '#/components/responses/DecodeFailed' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }

  /encode:
    post:
      summary: Encode a path as an OpenLR reference
      description: |
        `path` is an ordered list of `meta` values — the same references `/decode`
        returns. Prefix a value with `-` to traverse that segment against its
        digitised direction; only a *leading* `-` is a direction marker, so hyphens
        inside a value (a UUID, say) are part of the identifier.

        A `meta` value must identify exactly one segment. One matching several is
        rejected rather than resolved arbitrarily.
      operationId: encode
      tags: [Locations]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: array
                  items: { type: string }
                  example: ["390249024", "459554506", "-580286743"]
                positiveOffset:
                  type: integer
                  default: 0
                  description: Metres from the start of the path
                negativeOffset:
                  type: integer
                  default: 0
                  description: Metres from the end of the path
                props:
                  type: string
                  default: default
                  description: Encoder profile from `encoding_properties_dir`
      responses:
        '200':
          description: Encoded successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EncodeResponse' }
              example:
                success: true
                openLrCode: CwQ1cyVM2BNWDPvw/0cTBA==
                error: null
        '400':
          description: |
            Unresolvable or ambiguous `meta`, an impermissible direction, an unknown
            encoder profile, or an empty path element.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EncodeResponse' }
              examples:
                unknownMeta:
                  value:
                    success: false
                    openLrCode: null
                    error: "No segment found with meta 'no-such-segment'"
                ambiguous:
                  value:
                    success: false
                    openLrCode: null
                    error: "meta 'dup' matches 2 segments; encoding requires it to identify exactly one"
                wrongDirection:
                  value:
                    success: false
                    openLrCode: null
                    error: "Segment '123' cannot be traversed against its digitised direction (flowdir=START_TO_END)"
        '503': { $ref: '#/components/responses/DependencyUnavailable' }

  /roads/near:
    get:
      summary: Find segments near a coordinate
      operationId: roadsNear
      tags: [Network]
      parameters:
        - { name: lon, in: query, required: true, schema: { type: number, format: double }, example: 4.9 }
        - { name: lat, in: query, required: true, schema: { type: number, format: double }, example: 52.37 }
        - name: distance
          in: query
          required: false
          schema: { type: integer, default: 100 }
          description: Radius in metres
      responses:
        '200':
          description: Segments within the radius, nearest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  roads:
                    type: array
                    items: { $ref: '#/components/schemas/RoadSummary' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }

  /roads:
    get:
      summary: Find segments by meta value
      description: Exact match, no pattern matching.
      operationId: roadsByMeta
      tags: [Network]
      parameters:
        - { name: meta, in: query, required: true, schema: { type: string }, example: "390249024" }
      responses:
        '200':
          description: A GeoJSON FeatureCollection of the matching segments
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoadFeatureCollection' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }

  /roads/{id}:
    get:
      summary: Get a segment by internal ID
      description: |
        A diagnostic. `id` is the opaque `local.roads.id`; callers normally identify a
        segment by `meta` via `GET /roads`. The response body does not repeat the ID.
      operationId: roadById
      tags: [Network]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: The segment as a GeoJSON Feature
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoadFeature' }
        '404': { description: No such segment }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }

  /stats:
    get:
      summary: Row counts for the loaded network
      operationId: stats
      tags: [Operations]
      responses:
        '200':
          description: Counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  roadCount: { type: integer, example: 4717149 }
                  intersectionCount: { type: integer, example: 4329259 }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }

  /health:
    get:
      summary: Liveness and database connectivity
      description: |
        Touches the database, so it is a dependency check rather than liveness only.
        It runs two `COUNT(*)` queries — about 0.45 s on a 4.7M-road network.

        This is what the container health check probes. A probe must not mutate
        state: probing `POST /cache/clear` previously emptied every cache on each
        30-second interval.
      operationId: health
      tags: [Operations]
      responses:
        '200':
          description: Up, with the network loaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: UP }
                  roadCount: { type: integer }
                  intersectionCount: { type: integer }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }

  /cache/clear:
    post:
      summary: Empty the map caches
      description: |
        Idempotent. Needed after reloading map data, since the caches assume the
        network is immutable while the process runs.
      operationId: clearCache
      tags: [Operations]
      responses:
        '200':
          description: Cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: Caches cleared successfully }

  /cache/stats:
    get:
      summary: Cache occupancy, bounds and hit rates
      description: |
        `size` should settle at or below `maxSize`; a rising `evictions` means the
        bound is doing its job. Keys are `roads`, `intersections`, `lines`, `nodes`.
      operationId: cacheStats
      tags: [Operations]
      responses:
        '200':
          description: Statistics per cache
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { $ref: '#/components/schemas/CacheStats' }
              example:
                roads: { size: 9810, maxSize: 1000000, hits: 19232, misses: 46793, evictions: 0, hitRate: 0.29 }

  /properties/reload:
    post:
      summary: Re-read the decoder and encoder property files
      description: |
        Evicts the cached parameter sets so each profile's `.properties` file is read
        again on next use. Lets a profile be edited without a restart.
      operationId: reloadProperties
      tags: [Operations]
      responses:
        '200':
          description: Reloaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: Properties reloaded successfully }

components:
  parameters:
    OpenLrCode:
      name: openLrCode
      in: query
      required: true
      schema: { type: string }
      description: Base64-encoded binary OpenLR reference
      example: CwQ1cyVM2BNWDPvw/0cTBA==
    DecoderProps:
      name: props
      in: query
      required: false
      schema: { type: string, default: default }
      description: |
        Decoder profile, or a comma-separated list tried in order until one succeeds.
        Names must exist in `decoder_properties_dir`; an unknown name is rejected
        rather than silently falling back to the library defaults.
      example: strict,relaxed,default

  responses:
    DecodedPath:
      description: A GeoJSON FeatureCollection, one feature per matched segment, in path order
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DecodedFeatureCollection' }
    DecodeFailed:
      description: The reference could not be decoded, or a parameter was invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              msg: { type: string, example: "Failed to decode CwQ1cyVM2BNWDPvw/0cTBA==" }
              reason: { type: string }
          examples:
            noMatch:
              value:
                msg: "Failed to decode CwQ1cyVM2BNWDPvw/0cTBA=="
                reason: "Decoding failed with all property sets: default"
            unknownProfile:
              value:
                msg: "Failed to decode CwQ1cyVM2BNWDPvw/0cTBA=="
                reason: "Unknown decoder profile(s): typo. Available: default, relaxed, strict"
            unsupportedType:
              value:
                msg: "Failed to decode CwQ1cyVM2BNWDPvw/0cTBA=="
                reason: "Unsupported location type: POINT_ALONG_LINE"
    DependencyUnavailable:
      description: The map database is unavailable
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: Service Unavailable }
              reason: { type: string, example: "The map database is unavailable: Failed to obtain JDBC Connection" }

  schemas:
    LineStringGeometry:
      type: object
      properties:
        type: { type: string, enum: [LineString] }
        coordinates:
          type: array
          description: WGS84 [longitude, latitude] pairs
          items:
            type: array
            items: { type: number, format: double }
            minItems: 2
            maxItems: 2

    DecodedFeatureCollection:
      type: object
      properties:
        type: { type: string, enum: [FeatureCollection] }
        features:
          type: array
          items:
            type: object
            properties:
              type: { type: string, enum: [Feature] }
              geometry: { $ref: '#/components/schemas/LineStringGeometry' }
              properties:
                type: object
                properties:
                  meta:
                    type: string
                    description: |
                      Reference into the source network, and the only segment
                      identifier the API exposes. Collecting these in feature order
                      gives a valid `path` for `POST /encode`.
                  direction:
                    type: boolean
                    description: True when traversed with the digitised direction
                  frc: { type: string, example: FRC_3 }
                  fow: { type: string, example: SINGLE_CARRIAGEWAY }
                  lengthMeters: { type: integer }
        meta:
          type: object
          properties:
            posOff: { type: integer, description: Positive offset in metres }
            negOff: { type: integer, description: Negative offset in metres }
            openLR: { type: string, description: The submitted code, echoed }
            propertySet:
              type: string
              description: The profile that actually succeeded, when a list was given
            decodingElapsedNanos: { type: integer, format: int64 }
            wkt:
              type: string
              description: |
                The concatenated path. `LINESTRING EMPTY` when no feature resolved.
              example: LINESTRING(4.9 52.37, 4.901 52.371)

    RoadFeature:
      type: object
      properties:
        type: { type: string, enum: [Feature] }
        geometry: { $ref: '#/components/schemas/LineStringGeometry' }
        properties: { $ref: '#/components/schemas/RoadFeatureProperties' }

    RoadFeatureProperties:
      type: object
      properties:
        meta: { type: string }
        frc: { type: string, example: FRC_2 }
        fow: { type: string, example: ROUNDABOUT }
        flowDirection:
          type: string
          enum: [BOTH_WAYS, START_TO_END, END_TO_START]
          description: |
            Traversability. Derived from `local.roads.flowdir`: 1 two-way,
            2 one-way against the digitised direction, 3 one-way with it.
        lengthMeters: { type: number, format: double }
        startNodeId: { type: integer, format: int64 }
        endNodeId: { type: integer, format: int64 }

    RoadFeatureCollection:
      type: object
      properties:
        type: { type: string, enum: [FeatureCollection] }
        features:
          type: array
          items: { $ref: '#/components/schemas/RoadFeature' }
        meta:
          type: object
          properties:
            count: { type: integer }

    RoadSummary:
      type: object
      description: Attributes only; `/roads/near` returns no geometry
      properties:
        id:
          type: integer
          format: int64
          description: Internal key, included here for diagnostics only
        meta: { type: string, nullable: true }
        frc: { type: string }
        fow: { type: string }
        lengthMeters: { type: number, format: double }
        startNodeId: { type: integer, format: int64 }
        endNodeId: { type: integer, format: int64 }

    EncodeResponse:
      type: object
      properties:
        success: { type: boolean }
        openLrCode: { type: string, nullable: true }
        error: { type: string, nullable: true }

    CacheStats:
      type: object
      properties:
        size: { type: integer }
        maxSize: { type: integer }
        hits: { type: integer, format: int64 }
        misses: { type: integer, format: int64 }
        evictions: { type: integer, format: int64 }
        hitRate: { type: number, format: double, nullable: true }